feat: [SC-24139] Renamed core Avatar functions and separated core Avatar dependencies into 2 files
