+++ lib/client/avatar.ts
@@ -1,50 +1,4 @@
-import { Profile } from "@/data/ensProfiles";
-
-// Local Avatar utils
-const getCachedAvatarCallback: AvatarQueryModel = async (
-  addressOrEnsName: string
-) => {
-  return fetch(`/images/avatars/${addressOrEnsName}.png`)
-    .then((res) => {
-      if (res.ok) {
-        return res;
-      } else {
-        throw new Error(
-          `Failed to fetch cached avatar for ${addressOrEnsName}`
-        );
-      }
-    })
-    .catch((error) => {
-      throw new Error(error);
-    });
-};
-
-const getFallbackAvatarCallback: AvatarQueryModel = async () => {
-  return fetch("/images/no-avatar.png")
-    .then((res) => {
-      if (res.ok) {
-        return res;
-      } else {
-        throw new Error(`Failed to fetch fallback avatar`);
-      }
-    })
-    .catch((error) => {
-      throw new Error(error);
-    });
-};
-
-export const getNameKitAvatarCallbacks = (
-  profile: Profile
-): AvatarQueryModel[] => {
-  return [
-    async () => await getCachedAvatarCallback(profile.ensName),
-    async () => await getENSAvatar(profile.ensName),
-    async () => await getFallbackAvatarCallback(),
-  ];
-};
-
-// NameKit's library Avatar utils
-export const getENSAvatar: AvatarQueryModel = async (
+export const getDynamicENSAvatarCallback: AvatarQueryModel = async (
   addressOrEnsName: string
 ) => {
   return fetch(
